coll: fix 32-bit overflow in large-count segmentation arithmetic
CodeQL flagged int*int multiplications converted to size_t in several
collective components (alerts 68, 69, 70, 71, 72, 73, 74).  Most of
them are real bugs in the large-count (MPI-4 bigcount) paths:

  * coll/base bcast: the last-segment count is computed as
    original_count - segindex * count_by_segment.  The product is
    32-bit while original_count is a size_t user count, so any count
    requiring the product to exceed 32 bits wraps and produces a
    bogus final segment count.

  * coll/han allreduce, bcast, and reduce compute the same
    last-segment expression, count - (num_segments - 1) * seg_count,
    with the additional twist that the 32-bit product is signed, so
    overflow is undefined behavior.

  * coll/han scatter's local count variable was an int assigned from
    the size_t user counts -- and, on the reordered path, from a
    size_t per-rank *byte* count -- truncating multi-gigabyte values
    long before the flagged multiplication.  The variable is now
    size_t, which fixes both the truncation and the flagged product.

Compute all of these in size_t.  Two remaining sites are bounded well
below overflow (adapt ireduce's product is guarded by an
already-widened comparison one line earlier; han topo multiplies the
communicator size by 2) but are widened as well for consistency and
to resolve the alerts.

+++ ompi/mca/coll/han/coll_han_topo.c
@@ -199,7 +199,7 @@ mca_coll_han_topo_init(struct ompi_communicator_t *comm,
     }
 
     /* broadcast topology from node leaders to remaining ranks */
-    low_comm->c_coll->coll_bcast(topo, num_topo_level*size, MPI_INT, 0,
+    low_comm->c_coll->coll_bcast(topo, (size_t) num_topo_level * size, MPI_INT, 0,
                                 low_comm, low_comm->c_coll->coll_bcast_module);
     free(my_low_rank_map);
     han_module->cached_topo = topo;
